Fix MR comments (nc-694)

+++ nodeconductor/cost_tracking/migrations/0002_price_list.py
@@ -31,7 +31,7 @@ class Migration(migrations.Migration):
                 ('id', models.AutoField(verbose_name='ID', serialize=False, auto_created=True, primary_key=True)),
                 ('name', models.CharField(max_length=150, verbose_name='name')),
                 ('uuid', uuidfield.fields.UUIDField(unique=True, max_length=32, editable=False, blank=True)),
-                ('value', models.FloatField(default=0)),
+                ('value', models.DecimalField(default=0, max_digits=16, decimal_places=8)),
                 ('units', models.CharField(max_length=30, blank=True)),
                 ('price_list', models.ForeignKey(related_name='items', to='cost_tracking.PriceList')),
             ],
